County level voting and demographics data.

Demographics sourced from census.gov in 2019. Voting records from https://github.com/tonmcg/US_County_Level_Election_Results_08-16, sourced from The Guardian (2012) and Townhall.com (2016)

Usage

county_data

Format

county_data A data frame with 3142 rows and 65 columns:

FIPS

Five digit Federal Information Processing Standards code that uniquely identifies counties and county equivalents in the United States

state

State name

county_state

County and state names for display

state_abbrv

State abbreviation

county

County name

pop_estimate16

Population in the county in 2016

any_college

Percent of the county that attended college for some period of time, regardless of whether they got a degree.

college_2cat_num

Coded as 1 if the county is categorized as high college attendance, zero if low. See college_2cat for categorization rule.

college_2cat

High college attendance counties have over 51.24% college attendance (any_college). Low college attendance counties have under 51.24% college attendance. The mean value of any_college is 51.24.

prcnt_black

Percent of the county that is Black.

prcnt_unemployed

Percent of the population that is unemployed but looking for employment in 2016.

prcnt_unemployed_log

Logged percent of the population that is unemployed but looking for employment in 2016.

median_income16

o Median household income in the county in 2016.

median_income16_1k

Median household income in the county in 2016 in units of 1,000 dollars.

r_shift

Percentage difference between the Republican presidential vote in that county in 2016 and 2012. For example, 46.7955% of Kent County in Delaware (FIPS 20001) voted for Romney in 2012. In 2016, 49.81482% of that county voted for Trump. Therefore, the county shifted towards the Republican presidential candidate by 3.01325%. Positive value mean leaning more Republican; negative values mean leaning less Republican.

prcnt_GOP16

Percent of the county that voted for the Republican presidential candidate, Donald Trump, in 2016.

plurality_Trump16

Binary: 0 if less than a plurality of the county that voted for the Republican presidential candidate, Donald Trump, in 2016. 1 otherwise

Source

R/county_data.R