Add a line of predicted effects to the plotly surface, with optional confidence intervals. User can add a line in any direction.

Primarily used by functions such as add_3d_surface or add_marginals. If user defines "direction_data" appropriately, any line can be shown.

Usage

add_direction(
  p,
  model,
  direction_data,
  direction_name = "User defined line",
  linecolor = "black",
  ci = T
)

Arguments

p

A plotly object

model

A glm with exactly two x variables

direction_data

A data frame with of x values, predicted y values and predicted confidence interval for each pair of x values. The variable names are c("rownum", actual x1 variable name, actual x2 variable name, actual y variable name, "lowerCI", "upperCI")

direction_name

The hover text for the plotted line(s). Defaults to "User defined line".

linecolor

The color for the plotted line. Defaults to "black".

ci

An optional logical. Defaults to TRUE, showing the confidence intervals of the predicted effects.

Value

A plotly object